detector/requirements.txt
# Core
PyYAML==6.0.1
requests==2.31.0

# Web dashboard
Flask==3.0.0

# System monitoring (CPU/memory for dashboard)
psutil==5.9.8
